refactor create_person into CreatePerson, CreateTeacher and CreateStudent classes
