+++ app.rb
@@ -1,6 +1,6 @@
 require './validations'
 require './create_book'
-require './person'
+require './create_person'
 require './student'
 require './teacher'
 require './rental'
@@ -13,7 +13,7 @@ class App
     @people = []
     @books = []
     @create_book = CreateBook.new(@books)
-    @person = nil
+    @create_person = CreatePerson.new(@people)
     @rents = []
   end
 
@@ -24,7 +24,7 @@ class App
     when 2
       list_people
     when 3
-      create_person
+      @create_person.create_person
     when 4
       @create_book.create_book
     when 5
@@ -44,49 +44,6 @@ class App
     @people.each_with_index do |person, index|
       puts "#{index}) [#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
-  end
-
-  def create_person
-    loop do
-      print "Do you want to create a student (1) or a teacher (2)? [Input the number]:\s"
-      person = gets.chomp
-      case person
-      when '1'
-        create_student
-      when '2'
-        create_teacher
-      end
-
-      break if %w[1 2].include?(person)
-    end
-
-    puts "Person created successfully\n\n"
-  end
-
-  def create_student
-    age = numeric(message: "Age:\s")
-    name = not_empty(message: "Name:\s")
-    loop do
-      print "Has parent permission [Y/N]?\s"
-      permission = gets.chomp
-      if %w[y Y].include?(permission)
-        student = Student.new(nil, age, name: name, parent_permission: true)
-        @people << student unless @people.include?(student)
-        break
-      elsif %w[n N].include?(permission)
-        student = Student.new(nil, age, name: name, parent_permission: false)
-        @people << student unless @people.include?(student)
-        break
-      end
-    end
-  end
-
-  def create_teacher
-    age = numeric(message: "Age:\s")
-    name = not_empty(message: "Name:\s")
-    specialization = not_empty(message: "Specialization:\s")
-    teacher = Teacher.new(specialization, age, name: name)
-    @people << teacher unless @people.include?(teacher)
   end
 
   def create_rental
